refactor: 将用户、船只、船票等实体的 ID 类型从 Integer 修改为 Long，以支持更大的数据范围

+++ src/main/java/com/sakurapuare/boatmanagement/pojo/entity/Alert.java
@@ -23,7 +23,7 @@ import java.sql.Timestamp;
 public class Alert extends BaseEntity {
 
     @Id(keyType = KeyType.Auto)
-    private Integer alertId;
+    private Long alertId;
 
     @RelationOneToOne(selfField = "boatId", targetField = "boatId")
     private Boat boat;
